Update RabbitMQ User
The default RabbitMQ guest user cannot access the queue outside of localhost.
The platform now creates a "codelingo" user to access the queue. Update the
platform config template to create a codelingo user instead of guest. This
will enable the client to work with the remote production server as well as
the local dev one.