Async Progress Bar
Show a progress bar to indicate the ingest progress. If the client hangs up mid ingest and picks back up again, the progress bar picks up from where the ingest is up to.

This PR is a dependency of a platform PR. This PR needs to land first so the glide.yaml SHA can be updated.