Tidy Up Repo Sync
Fix a bug where the local and remote were out of sync, causing a new remote repository to be created on every review without updating. Remove the redundant InitRepo method and "lazy load" the repository on each sync attempt:

- Attempt to sync the local repo with the remote on each review
- On failure, create a new remote, add the new remote address to git and attempt to sync again

This PR also improves an error message to prompt user to add first commit if missing.